# internal/interfaces/rest/openapi.yaml
openapi: 3.1.0
info:
  title: Authentication Service API
  description: |
    AuthService is a self-hosted, multi-tenant authentication platform for teams that want provider-grade auth without handing user data, pricing leverage, or service-to-service trust to a third party.

    It provides email/password auth, OAuth2 social login, enterprise SSO with SAML/OIDC, SCIM 2.0 directory sync, machine-to-machine client credentials, magic links, TOTP 2FA, adaptive security and step-up, passkeys/WebAuthn, organization RBAC, refresh-token rotation, tenant-scoped JWTs, JWKS, audit logs, REST, gRPC, and an importable Go JWT validator.

    ## What This Replaces

    Use AuthService when you would otherwise reach for Auth0, Clerk, Firebase Auth, Supabase Auth, Cognito, Stytch, WorkOS, or Keycloak but need a lightweight service your company can own, deploy, inspect, and extend.

    | Need | AuthService Capability |
    |------|------------------------|
    | Product authentication | Signup, login, refresh, logout, profile, password change |
    | Passwordless login | Magic links and passkeys/WebAuthn |
    | MFA | TOTP setup, enable, login verify, and disable lifecycle |
    | Adaptive security | Risk-driven MFA, remembered devices, admin/user step-up, and policy-driven challenge/block/notify decisions |
    | Social identity | Google, GitHub, Microsoft, and Apple OAuth2 |
    | Enterprise SSO | Per-client SAML 2.0 and OIDC connections, domain routing, metadata, and JIT provisioning |
...

...

...
servers:
...
paths:
...
  /api/auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh an access token
      description: |
        Exchanges a refresh token for a new access token. Refresh tokens are single-use and rotated on each call.

        Provide the refresh token either as:
        - `refresh_token` in the request body (token mode)
        - `auth_refresh` cookie (cookie mode)
      tags: [Authentication]
      x-codeSamples:
        - lang: cURL
          label: Rotate a token-mode refresh token
          source: |
            curl -X POST https://authservice.ayushojha.com/api/auth/refresh \
              -H "X-API-Key: $AUTH_SERVICE_API_KEY" \
              -H "Content-Type: application/json" \
              -d '{
                "refresh_token": "$AUTH_REFRESH_TOKEN",
                "session_mode": "token"
              }'
      security:
        - ApiKeyAuth: []
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshRequest"
            examples:
              cookieMode:
                summary: Browser refresh using auth_refresh cookie
                value: {}
              tokenMode:
                summary: Token refresh for native, CLI, or API integrations
                value:
                  refresh_token: rfr_opaque_refresh_token
                  session_mode: token
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid or expired refresh token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/auth/sso:
    get:
      operationId: beginEnterpriseSSOByDomain
      summary: Start enterprise SSO by email domain
      description: Resolves an active SAML or OIDC connection by domain and redirects the browser to the enterprise identity provider.
      tags: [Enterprise SSO]
      security:
        - ApiKeyAuth: []
      parameters:
        - name: domain
          in: query
          required: true
          schema:
            type: string
          example: acme.com
        - name: session_mode
          in: query
          required: false
          schema:
            type: string
            enum: [token]
      responses:
        "302":
          description: Redirect to the configured SAML or OIDC identity provider
        "404":
          description: No active SSO connection is configured for the domain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
  /api/auth/sso/{sso_connection_id}:
    get:
      operationId: beginEnterpriseSSOByConnection
      summary: Start enterprise SSO by connection
      description: Starts SSO using a connection ID or slug and redirects the browser to the enterprise identity provider.
      tags: [Enterprise SSO]
      security:
        - ApiKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/SSOConnectionID"
        - name: session_mode
          in: query
          required: false
          schema:
            type: string
            enum: [token]
      responses:
        "302":
          description: Redirect to the configured SAML or OIDC identity provider
        "400":
          description: Invalid or inactive SSO connection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
  /api/admin/auth/sso:
    post:
      operationId: adminSSOLogin
      summary: Sign in an admin user with SSO
      description: Completes an externally verified admin SSO handoff by matching provider and subject to an admin identity.
      tags: [Admin]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [provider, subject]
              properties:
                provider:
                  type: string
                subject:
                  type: string
      responses:
        "200":
          description: Admin token
...
  /api/admin/clients/{id}:
    get:
      operationId: getClient
      summary: Get a client by ID
      description: Returns a single client's details.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Client details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
    patch:
      operationId: updateClient
      summary: Update a client
      description: Updates mutable client settings such as allowed origins, webhook URL, status, and tenant WebAuthn policy controls.
      tags: [Admin]
      security:
        - AdminBearerAuth: []
        - AdminKeyAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientRequest"
            example:
              settings:
                webauthn_attestation: direct
                webauthn_require_attestation: true
                webauthn_allowed_attestation_formats: [packed, tpm, apple]
      responses:
        "200":
          description: Client updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "401":
          description: Invalid admin key
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "404":
          description: Client not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: Client API key — identifies the tenant
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User access token (JWT)
    SCIMBearerAuth:
      type: http
      scheme: bearer
      description: Per-directory SCIM provisioning token
    AdminKeyAuth:
      type: apiKey
      in: header
      name: X-Admin-Key
      description: Break-glass master admin key for bootstrap or emergencies; prefer AdminBearerAuth
    AdminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin access token returned by `/api/admin/auth/login` or `/api/admin/auth/sso`
  parameters:
...
    SSOConnectionID:
      name: sso_connection_id
      in: path
      required: true
      schema:
        type: string
      description: Enterprise SSO connection ID or slug where supported
...
  schemas:
...
    RefreshRequest:
      type: object
      properties:
        refresh_token:
          type: string
          description: Refresh token (required in token mode)
        session_mode:
          type: string
          enum: [token]
...
    UpdateClientRequest:
      type: object
      properties:
        name:
          type: string
        allowed_origins:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          format: uri
        status:
          type: string
          enum: [active, suspended]
        settings:
          type: object
          additionalProperties: true
          description: Replaces the client settings object when supplied.

    # ── Responses ─────────────────────────────
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
          example: invalid request
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token (15 min TTL)
        token_type:
          type: string
          example: Bearer
        expires_in:
          type: integer
          example: 900
          description: Token lifetime in seconds
        user:
          $ref: "#/components/schemas/User"
        refresh_token:
          type: string
          description: Refresh token (only present in token session mode)
        risk:
          $ref: "#/components/schemas/LoginRisk"
...
    LoginRisk:
      type: object
      description: Present when a login is considered suspicious and should be surfaced to the user or security team.
      properties:
        level:
          type: string
          enum: [low, medium, high, critical]
          example: medium
        score:
          type: integer
          minimum: 0
          maximum: 100
        reasons:
          type: array
          items:
            type: string
          example: [new_ip, new_device, failed_velocity]
        new_ip:
          type: boolean
        new_device:
          type: boolean
        signals:
          type: array
          items:
            $ref: "#/components/schemas/RiskSignal"
        context:
          type: object
          additionalProperties: true
    RiskSignal:
      type: object
      properties:
        type:
          type: string
          example: new_device
        level:
          type: string
          enum: [low, medium, high, critical]
        reason:
          type: string
        metadata:
          type: object
          additionalProperties: true
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        client_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        email_verified:
          type: boolean
        display_name:
          type: string
        avatar_url:
          type: string
          nullable: true
        timezone:
          type: string
          example: UTC
        locale:
          type: string
          example: en
        role:
          type: string
          example: user
        status:
          type: string
          enum: [active, suspended, deleted]
        totp_enabled:
          type: boolean
        last_login_at:
          type: string
          format: date-time
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Client:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        slug:
          type: string
        allowed_origins:
          type: array
          items:
            type: string
        webhook_url:
          type: string
          nullable: true
          description: Receives signed `audit.event` webhooks when webhook delivery is enabled
        settings:
          type: object
        status:
          type: string
          enum: [active, suspended]
        token_mode:
          type: string
          enum: [v1_hs256, v2_jwks]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...